ModLogMirror/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
charset-normalizer==3.1.0
discord.py==2.3.0
distlib==0.3.6
filelock==3.12.2
frozenlist==1.3.3
idna==2.10
importlib-metadata==3.4.0
multidict==5.1.0
mysql-connector-python==8.0.33
Pillow==9.5.0
pip==23.1.2
pipenv==2023.6.12
platformdirs==3.5.3
praw==7.7.0
prawcore==2.3.0
protobuf==3.14.0
psycopg2==2.9.6
PyYAML==5.4.1
redis==4.5.5
requests==2.25.1
setuptools==67.8.0
six==1.15.0
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.2
virtualenv==20.23.0
virtualenv-clone==0.5.4
websocket-client==0.57.0
websockets==8.1
yarl==1.6.3
zipp==3.4.0
